Deprecate functions & update esQuery/$
### Added
- Add `ready`, `loaded`, `mediaQuery`, & `getLocation` static methods to `esQuery`

### Changed
- Move deprecated function in `functions.js` to separate script
- Set `esQuery` static methods on the `$` function when exporting
- Rename `wait()` to `sleep()` and alias old function with warning
- Misc. updates to reuse functions instead of repeating them
- Update linting & dependencies
