+++ test/index.css
@@ -5,6 +5,7 @@
 @import url("https://cdn.kernvalley.us/css/core-css/class-rules.css");
 @import url("https://cdn.kernvalley.us/css/core-css/scrollbar.css");
 @import url("https://cdn.kernvalley.us/css/core-css/forms.css");
+@import url("https://cdn.kernvalley.us/css/core-css/animations.css");
 @import url("https://cdn.kernvalley.us/css/core-css/layout/default/index.css");
 @import url("https://cdn.kernvalley.us/css/core-css/theme/properties.css");
 @import url("https://cdn.kernvalley.us/css/core-css/theme/adwaita/index.css");
